Refactor interrupt execution logic into shared hook
Extract duplicated execution interruption logic from AiCell, CodeCell, and SqlCell into a reusable useInterruptExecution hook. This reduces code duplication and provides a consistent interface with hasActiveExecution state.
