+++ src/components/notebook/cell/AiCell.tsx
@@ -5,9 +5,10 @@ import { useCellOutputs } from "@/hooks/useCellOutputs.js";
 import { useAuth } from "@/components/auth/AuthProvider.js";
 import { useUserRegistry } from "@/hooks/useUserRegistry.js";
 import { useToolApprovals } from "@/hooks/useToolApprovals.js";
+import { useInterruptExecution } from "@/hooks/useInterruptExecution.js";
 import { useAvailableAiModels } from "@/util/ai-models.js";
-import { queryDb } from "@livestore/livestore";
-import { useStore, useQuery } from "@livestore/react";
+
+import { useStore } from "@livestore/react";
 import { events, tables } from "@runt/schema";
 import { ChevronDown, ChevronUp } from "lucide-react";
 import React, { useCallback } from "react";
@@ -175,31 +176,11 @@ export const AiCell: React.FC<AiCellProps> = ({
     }
   }, [cell.id, localSource, cell.source, cell.executionCount, store, userId]);
 
-  // Query execution queue for this cell
-  const executionQueue = useQuery(
-    queryDb(tables.executionQueue.select().where({ cellId: cell.id }))
-  );
-
-  const interruptAiCell = useCallback(async () => {
-    // Find the current execution in the queue for this cell
-    const currentExecution = executionQueue.find(
-      (exec: any) =>
-        exec.status === "executing" ||
-        exec.status === "pending" ||
-        exec.status === "assigned"
-    );
-
-    if (currentExecution) {
-      store.commit(
-        events.executionCancelled({
-          queueId: currentExecution.id,
-          cellId: cell.id,
-          cancelledBy: userId,
-          reason: "User interrupted AI execution",
-        })
-      );
-    }
-  }, [cell.id, store, userId, executionQueue]);
+  const { interruptExecution: interruptAiCell } = useInterruptExecution({
+    cellId: cell.id,
+    userId,
+    reason: "User interrupted AI execution",
+  });
 
   // Use shared keyboard navigation hook
   const { keyMap } = useCellKeyboardNavigation({
