+++ src/components/notebook/cell/SqlCell.tsx
@@ -4,8 +4,9 @@ import { useCellKeyboardNavigation } from "@/hooks/useCellKeyboardNavigation.js"
 import { useCellOutputs } from "@/hooks/useCellOutputs.js";
 import { useAuth } from "@/components/auth/AuthProvider.js";
 import { useUserRegistry } from "@/hooks/useUserRegistry.js";
-import { queryDb } from "@livestore/livestore";
-import { useStore, useQuery } from "@livestore/react";
+import { useInterruptExecution } from "@/hooks/useInterruptExecution.js";
+
+import { useStore } from "@livestore/react";
 import { events, tables } from "@runt/schema";
 import { ChevronDown, ChevronUp } from "lucide-react";
 import React, { useCallback } from "react";
@@ -105,32 +106,11 @@ export const SqlCell: React.FC<SqlCellProps> = ({
     }
   }, [cell.id, store, hasOutputs, userId]);
 
-  // Query execution queue for this cell
-  const executionQueue = useQuery(
-    queryDb(tables.executionQueue.select().where({ cellId: cell.id }))
-  );
-
-  const interruptQuery = useCallback(() => {
-    // Find the current execution in the queue for this cell
-
-    const currentExecution = executionQueue.find(
-      (exec: any) =>
-        exec.status === "executing" ||
-        exec.status === "pending" ||
-        exec.status === "assigned"
-    );
-
-    if (currentExecution) {
-      store.commit(
-        events.executionCancelled({
-          queueId: currentExecution.id,
-          cellId: cell.id,
-          cancelledBy: userId,
-          reason: "User interrupted SQL execution",
-        })
-      );
-    }
-  }, [cell.id, store, userId, executionQueue]);
+  const { interruptExecution: interruptQuery } = useInterruptExecution({
+    cellId: cell.id,
+    userId,
+    reason: "User interrupted SQL execution",
+  });
 
   // Use shared keyboard navigation hook
   const { keyMap } = useCellKeyboardNavigation({
